Implement freshness status flag.

+++ MiseTrack/Packages/Base/Models/Sources/Models/Config/BatchLimits.swift
@@ -7,14 +7,14 @@
 
 public struct BatchLimits: Codable, Sendable {
     public let batchAmountLimitMl: Double
-    public let batchExpirationInSeconds: Int
+    public let batchExpirationInSeconds: Double
     
     enum CodingKeys: String, CodingKey {
         case batchAmountLimitMl = "batch_amount_limit_ml"
         case batchExpirationInSeconds = "batch_expiration_in_seconds"
     }
     
-    public init(batchAmountLimitMl: Double, batchExpirationInSeconds: Int) {
+    public init(batchAmountLimitMl: Double, batchExpirationInSeconds: Double) {
         self.batchAmountLimitMl = batchAmountLimitMl
         self.batchExpirationInSeconds = batchExpirationInSeconds
     }
